test: add route tests and split DB tests into test_database.py
- rename test_app.py -> test_database.py (DB-layer tests); add a fresh
  test_app.py for the FastAPI route tests
- add conftest.py with the shared session fixture (isolated in-memory engine,
  StaticPool + check_same_thread for cross-thread TestClient) plus
  race_events / standings_pairs / seeded_season
- cover /, /standings/drivers, /standings/constructors, /results/{circuit_id},
  and /replay's 404 via TestClient, overriding get_session onto the test
  session with the lifespan left dormant
- document why test_championship_entry_link builds the Championship by hand
  rather than via create_championship (keeps the ORM-relationship test
  independent of the mutator)

Refs #65
